==> 1 Property Market Scraping - Main Page.ipynb <==
# Scraping
## Property Market

# Adjust Notebook Display
from IPython.display import display, HTML
display(HTML("<style>.container { width:90% !important; }</style>"))

# Imports

# Generic Web Scrapers
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.by import By

# Generic Packages
import re
import time
import pandas as pd
import numpy as np

# Chrome Options
options = webdriver.ChromeOptions()
options.add_argument("start-maximized") 
options.add_experimental_option("excludeSwitches", ["enable-automation"]) 
options.add_experimental_option('useAutomationExtension', False)
options.add_argument("--incognito")
options.add_argument("--disable-gpu")

# Locate Chrome Driver
service_driver = Service('chromedriver.exe')

# Add Customisation to Driver
driver = webdriver.Chrome(service=service_driver, options=options)

### Scrape Main Listing Pages

# Sraper

price = []
links = []
beds = []
property_type = []
location = []

counter = 1
last_page = 658

while counter <= last_page:
    
    # Create url
    url = 'https://www.propertymarket.com.mt/for-sale/?li=&mnp=0&mxp=0&pc=-1&pt=0&nb=0&o=1&d=5&f=&pp=' + str(counter)
    
    # Display Progress
    print(url, 'Page', counter, 'out of', last_page)
    
    # Open Page
    try:
        driver.get(url)
    except:
        print('Issue with Opening New Page, Scraping Stopped')
        break
    
    # Prices
    for i in driver.find_elements(By.CLASS_NAME, "searchResultListingPrice "):
        if i.text in ["Can't find what you are looking for?", "Thinking of selling your property?"]:
            pass
        else:
            price.append(i.text.replace('€', '').replace(',', ''))
    # Get Description Links
            links.append(i.find_element(By.TAG_NAME, "a").get_attribute('href'))
        
    # Title
    for i in driver.find_elements(By.CLASS_NAME, "searchResultListingTitle"):
    # Beds
        try:
            beds.append(int(i.text[0]))
        except:
            beds.append(np.nan)
    # Property Type
        try:
            property_type.append(re.search('Bedroom (.*) For', i.text).group(1))
        except:
            property_type.append(i.text[0])
    # Location
        location.append(re.search('in (.*)', i.text).group(1))
        
    # Check for inconsistencies
    if len(price) != len(beds):
        print('Issue with Scraping Found')
        print('Scraper Stopped')
        break
        
    # Wait before moving to next page
    time.sleep(3 + np.random.choice([x/10 for x in range(7,22)]))
        
    # Move to Next Page
    counter += 1
    
# Completed Scraping
driver.close()

# Get scraped data into dataframe format

# Create empty dataframe
prop_market = pd.DataFrame()

# Populate empty dataframe with scraped data
prop_market['Links'] = links
prop_market['Locality'] = location
prop_market['Price'] = price
prop_market['Beds'] = beds
prop_market['Property Type'] = property_type

# View dataframe
print('DataFrame Size :', prop_market.shape)
prop_market.head()

For the task of residential property valuation we are not interested in commercial properties and plots of land for sale. Such observations have the Beds column with NAs. Hence, by dropping NAs these observations are removed.

# Drop NAs
prop_market = prop_market.dropna()

Since we are concern with property valuations, observations with prices available on request are not useful.

# Remove observations without property price
prop_market = prop_market[prop_market['Price'] != "Price on Request"]

# Save Data
prop_market.to_csv('prop_market.csv', index=False)
