==> 3 Property Market - Data Cleaning.ipynb <==
# Data Cleaning
## Property Market

# General Data Analysis
import pandas as pd
import numpy as np

# Text Based Libraries
import re
import nltk
import gensim
from nltk.stem import SnowballStemmer   
from nltk.stem.wordnet import WordNetLemmatizer
from gensim.parsing.preprocessing import STOPWORDS
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import TfidfVectorizer

# Import Created Custom Functions from other files
from ipynb.fs.full.TextFunctions import abbr_or_lower, tokenize, find_words, get_feature

# Load Scraped Data
prop_market = pd.read_csv('prop_market.csv')
print('Property Market Dataframe Size :', prop_market.shape)
prop_market.head()

# Quick Check of Nulls
prop_market.info()

# Drop Rows with Nulls in Baths and Description Columns
prop_market = prop_market.dropna(subset = ['Baths', 'Description'])

# Re-Check Status of Nulls
prop_market.info()

# Drop Size and Features Column - Excessive Portion of Nulls
prop_market = prop_market.drop(['Size', 'Features'], axis=1)

# Get Baths as Integer
prop_market['Baths'] = prop_market['Baths'].astype(int)

# Convert Price Column Type for better Analysis

# Correct for Price Ranges
prop_market = prop_market.reset_index()
prop_market = prop_market.drop(['index'], axis=1)
for i in range(len(prop_market)):
    if len(prop_market.loc[i, 'Price']) > 8:
        prop_market.loc[i, 'Price'] = (int(re.search('(.*) ', prop_market.loc[i, 'Price']).group(1)) + int(re.search(' (.*)', prop_market.loc[i, 'Price']).group(1)))/2
        
prop_market['Price'] = prop_market['Price'].astype(int)

# Quick check of dataframe
prop_market.describe(include = ['object', 'float', 'int'])

Notes:
* Number of unique localities is more than the number of local councils in Malta & Gozo (68)
* Prices of 1 are usually alternatives to writing 'Price on Request' and prices less than 2,000 EUR are assumed incorrect
* Large Variety of Property Types
* Description requires text cleaning

#### Locality

prop_market['Locality'].value_counts()

Low observation counts for hamlets and high-rise developments. This can lead to statitical insignificance in modelling.

# Show included localities
prop_market['Locality'].unique()

# Assign Hamlets and Highrises to respective Local Councils

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Guardamangia'), 'Pieta', prop_market['Locality'])

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Zebbiegh'), 'Mgarr', prop_market['Locality'])

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Gozo - Santa Lucija'), 'Gozo - Kercem', prop_market['Locality'])

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Gozo - Marsalforn'), 'Gozo - Zebbug', prop_market['Locality'])

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Gozo - Mgarr'), 'Gozo - Ghajnsielem', prop_market['Locality'])

prop_market['Locality'] = np.where((prop_market['Locality'] == 'Gozo - Xlendi'), 'Gozo - Munxar', prop_market['Locality'])

sgn = ['Kappara', 'Mensija']
for i in sgn:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'San Gwann', prop_market['Locality'])

bkr = ['Swatar', 'Fleur-de-Lys']
for i in bkr:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Birkirkara', prop_market['Locality'])

slm = ['Tigne Point', 'Fort Cambridge']
for i in slm:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Sliema', prop_market['Locality'])

rbt = ['Bahrija', 'Bingemma']
for i in rbt:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Rabat', prop_market['Locality'])

stj = ['Portomaso', 'Pendergardens', 'Paceville', "Ta' Giorni", "The Gardens (St Julian's)"]
for i in stj:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), "St Julian's", prop_market['Locality'])
    
nxr = ['Maghtab', 'Bahar Ic-Caghaq', 'Birguma', 'San Pawl tat Targa', 'Salina']
for i in nxr:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Naxxar', prop_market['Locality'])
    
mlh = ['Santa Maria Estate (Mellieha)', 'Manikata']
for i in mlh:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Mellieha', prop_market['Locality'])
                                       
spb = ['Bugibba', 'Xemxija', 'Qawra', 'Burmarrad', 'Wardija']
for i in spb:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), "St Paul's Bay", prop_market['Locality'])
    
swq = ['Madliena', "Ta' l-Ibragg", 'High Ridge']
for i in swq:
    prop_market['Locality'] = np.where((prop_market['Locality'] == i), 'Swieqi', prop_market['Locality'])
    
print('Number of Localities :', len(prop_market['Locality'].unique().tolist()))

#### Price

# Remove Prices below 2000 
prop_market = prop_market[prop_market['Price'] > 2000]
print('Dataframe Size :', prop_market.shape)

#### Property Types

# Show included property types
prop_market['Property Type'].unique()

# Adjust Scraping Errors

prop_market['Property Type'] = np.where((prop_market['Property Type'] == 'Apartment For Sale in'), 'Apartment', prop_market['Property Type'])

prop_market['Property Type'] = np.where((prop_market['Property Type'] == 'Penthouse For Sale in'), 'Penthouse', prop_market['Property Type'])

prop_market['Property Type'] = np.where((prop_market['Property Type'] == 'Terraced House For Sale in'), 'Terraced House', prop_market['Property Type'])

#### Description

# Remove Common Start of Description Text
prop_market['Description'] = prop_market['Description'].str.replace('Property description', '')

# Clean Text with lowercase and lemmatization
prop_market['Description - Clean'] = [tokenize(listing, 2) for listing in prop_market['Description']]

## Create Features From Text

#### Garden

# Find words relating to gardens
find_words(prop_market['Description - Clean'], 'gard', 'in')

# Get Dummy for Properties with Gardens
garden_words = ['backgarden', 'garden', 'garden2', 'gardenfreeholde', 'gardenget', 'gardenia', 'gardening', 'gardenluxury',
                'gardensexperience', 'gardenstreet', 'gardenuca', 'terracegarden']
prop_market['Garden'] = get_feature(prop_market['Description - Clean'], garden_words)

prop_market['Garden'].value_counts()

#### Yard

# Find words relating to yards
find_words(prop_market['Description - Clean'], 'yard', 'in')

# Get Dummy for Properties with Yards
yard_words = find_words(prop_market['Description - Clean'], 'yard', 'in')
prop_market['Yard'] = get_feature(prop_market['Description - Clean'], yard_words)

prop_market['Yard'].value_counts()

#### Terrace

# Find words relating to terraces
find_words(prop_market['Description - Clean'], 'terrace', 'in')

# Get Dummy for Properties with Terraces
terrace_words = ['balconyterraceroof', 'bathroombackyardpoolbalconyterraceroof', 'bathroomsterrace', 
              'bathroomsterraceyardroof', 'bathroomterrace', 'bathroomterracebalconyroof', 'bathroomterracefreeholde', 
              'bathroomterraceroof', 'courtyardterracefreeholde', 'diningterraceone', 'edgeterraced', 'roomterrace',
              'terrace', 'terrace2', 'terraceasking', 'terraceback', 'terracebalconyfreeholde', 'terracebalconyserved', 
              'terracecall', 'terracefinished', 'terracefreeholde', 'terracegarden', 'terracehighly', 'terraceincluding',
              'terraceinternal', 'terraceoptional', 'terraceref', 'terraceroof', 'terraces3', 'terracesecond',
              'terraceserved', 'terracesfreeholde', 'terracestep', 'terracesthis', 'terracesupvc', 'terracethis']
prop_market['Terrace'] = get_feature(prop_market['Description - Clean'], terrace_words)

prop_market['Terrace'].value_counts()

#### Lift

# Find words relating to lifts
find_words(prop_market['Description - Clean'], 'lift', 'in')

# Get Dummy for Properties with Lifts
lift_words = ['garagelift', 'lift', 'liftairspace', 'liftfreehold', 'liftfreeholde', 'liftown', 'liftpartly', 'liftref',
              'liftroof']
prop_market['Lift'] = get_feature(prop_market['Description - Clean'], lift_words)

prop_market['Lift'].value_counts()

#### Pool

# Find words relating to pools
find_words(prop_market['Description - Clean'], 'pool', 'in')

# Get Dummy for Properties with Pools
pool_words = ['60sqmpool', 'bathroombackyardpoolbalconyterraceroof', 'bathroompool', 'pool', 'pool1', 'poolfreeholde',
                'poolget', 'poolref', 'poolside', 'poolsref', 'pooluca']
prop_market['Pool'] = get_feature(prop_market['Description - Clean'], pool_words)

prop_market['Pool'].value_counts()

#### Roof Access

# Find words relating to roofs
find_words(prop_market['Description - Clean'], 'roof', 'in')

# Get Dummy for Properties with Roof Access
roof_words = ['130sqmroof', 'areabalconyroof', 'backyardsroof', 'balconyroof', 'balconyterraceroof',
              'bathroombackyardpoolbalconyterraceroof', 'bathroombackyardroof', 'bathroombalconygarageroof',
              'bathroombalconyroof', 'bathroomcourtyardroof', 'bathroomroof', 'bathroomsbackyardroof',
              'bathroomsbalconyroof', 'bathroomsroof', 'bathroomsterraceyardroof', 'bathroomterracebalconyroof',
              'bathroomterraceroof', 'bedroomsroof', 'courtyardbalconiesroof', 'courtyardroof', 'floorwashroomroof',
              'garageroof', 'liftroof', 'maisonetteroof', 'roof', 'rooffreeholde', 'rooffreeholdoptional', 'roofonly',
              'roofprivate', 'roofref', 'roofsituated', 'roofthis', 'rooftop', 'roofucafreehold79026090ref', 'roomroof',
              'terraceroof', 'toproof', 'wardroberoof', 'washroomroof', 'yardbalconyroof', 'yardroof']
prop_market['Roof Access'] = get_feature(prop_market['Description - Clean'], roof_words)

prop_market['Roof Access'].value_counts()

#### Seafront

# Find words relating to seafront properties
find_words(prop_market['Description - Clean'], 'sea', 'in')

# Get Dummy for Properties at the Seafront
sea_words = ['brendaseafront', 'seafront', 'seafrontref', 'seafrontstructurally', 'seashore', 'seaside']
prop_market['Seafront'] = get_feature(prop_market['Description - Clean'], sea_words)

prop_market['Seafront'].value_counts()

#### Views

# Find words relating to properties with views
find_words(prop_market['Description - Clean'], 'view', 'in')

# Get Dummy for Properties with Views
view_words = ['kitchenviewopen', 'seaview', 'seaview3', 'seaviews', 'seaviewsopen', 'seaviewsref', 'views1',
              'viewsapartment', 'viewsasking', 'viewscall', 'viewscirca', 'viewsenjoy', 'viewsfinished', 'viewsfully',
              'viewsref']
prop_market['Views'] = get_feature(prop_market['Description - Clean'], view_words)

prop_market['Views'].value_counts()

#### Garage

# Find words relating to properties with garages
find_words(prop_market['Description - Clean'], 'gar', 'in')

# Get Dummy for Properties with Garages
garage_words = ['bathroombalconygarageroof', 'garage', 'garage315sqm', 'garageairspacefreeholde', 'garageasking',
                'garageavailable', 'garagecall', 'garagefinished', 'garagefreeholde', 'garageget', 'garageinterconnected',
                'garageit', 'garagelift', 'garagelovely', 'garagepermits', 'garagepossibility', 'garageref', 'garageroof',
                'garageserved', 'garagesfreeholde', 'garagesref', 'garagestudyreading', 'garagetotal', 'garageuca',
                'garges', 'maisonettesapartmentspenthousesgaragesre', 'torrimaisonettesapartmentspenthousesgaragesre']
prop_market['Garage'] = get_feature(prop_market['Description - Clean'], garage_words)

prop_market['Garage'].value_counts()

#### Garage Option

Using bigrams to see words that denote 'option' relating to gragaes in the descriptions.

# Instanciate TfidfVectorizer (tokenizer set to iterate over documents/paragraphs)
tfidf2 = TfidfVectorizer(ngram_range=(2, 2), tokenizer=lambda doc: doc, lowercase=False) # Bigrams

# Keep descriptions that references garages
garage_in_desc = []
for i in prop_market['Description - Clean']:
    if len(set(i).intersection(garage_words)) != 0:
        garage_in_desc.append(i)

# Bigrams
tfidf_description = tfidf2.fit_transform(garage_in_desc)
bigrams_df = pd.DataFrame.sparse.from_spmatrix(tfidf_description, columns=tfidf2.get_feature_names_out())
desc_bigrams = list(bigrams_df.columns)

# Get bigrams that reference garages
garage_bigrams = []
for i in desc_bigrams:
    if 'garage' in i:
        garage_bigrams.append(i)
        
print('Number of Bigrams Referencing Garages :', len(garage_bigrams))

# Manual identification of bigrams with references to garages on offer
garage_opt_bi = ['2023optional garage', '2024optional garage', '2025optional garage', '235koptional garage',
                 '30sqmoptional garagefreeholde', 'availability garage', 'available garage',
                 'bathroomsbalconybackyardoptional garageroof', 'buy garage', 'car garagepossibility',
                 'conditionedoptional garage', 'freeholdoptional garage', 'garage additional', 'garage avaialble',
                 'garage availabel', 'garage available', 'garage availablecontact', 'garage availableextras',
                 'garage availableget', 'garage availableproject', 'garage availableref', 'garage avilable',
                 'garage offer', 'garage offered', 'garage offering', 'garage opt', 'garage option', 'garage optional',
                 'garage optionalref', 'garage possibility', 'garage possible', 'garage possibly', 'garage potential',
                 'garage rent', 'garage rented', 'garage request', 'garage requested', 'garage315sqm asking',
                 'garageasking price', 'garageavailable asking', 'garagepossibility gym38', 'house garageasking',
                 'leasing garage', 'let garage', 'option garage', 'optional garage', 'optional garageavailable',
                 'optional garagefinished', 'optional garagesref', 'purchase garage', 'rent garage', 'rental garage',
                 'rented garage', 'streetoptional garage', 'terraceoptional garage']

# Words Denoting Options
opt_words = ['2023optional', '2024optional', '2025optional', '235koptional', '30sqmoptional', 'additional', 'asking',
             'avaialble', 'availabel', 'availability', 'available', 'availablecontact', 'availableextras', 'availableget',
             'availableproject', 'availableref', 'avilable', 'bathroomsbalconybackyardoptional', 'buy', 'rental', 'rented',
             'conditionedoptional', 'freeholdoptional', 'garageasking', 'garageavailable', 'garagepossibility', 'leasing',
             'let', 'offer', 'offered', 'offering', 'opt', 'option', 'optional', 'optionalref', 'possibility', 'possible',
             'possibly', 'potential', 'purchase', 'rent', 'request', 'requested', 'streetoptional', 'terraceoptional']

# Get dummy for optional garage
garage_opt_dummy = []
for i in prop_market['Description - Clean']:
    # Loop through features to check for 'garage' reference
    if len(set(i).intersection(garage_words)) == 0:
        garage_opt_dummy.append(0)
    else:
        # Check for reference of garage as option
        if len(set(i).intersection(opt_words)) == 0:
            garage_opt_dummy.append(0)
        else:
            garage_opt_dummy.append(1)

# Assign Dummy to Dataframe
prop_market['Optional Garage'] = garage_opt_dummy

prop_market['Optional Garage'].value_counts()

# Correct overlap between dummy for garage and optional garage
prop_market['Garage'] = np.where((prop_market['Optional Garage'] == 1), 0, prop_market['Garage'])
prop_market['Garage'].value_counts() #confirmed, correction worked

#### Airspace

# Find words relating to properties with airspace included
find_words(prop_market['Description - Clean'], 'airspace', 'in')

# Get Dummy for Properties with Airspace
airspace_words = find_words(prop_market['Description - Clean'], 'airspace', 'in')
prop_market['Airspace'] = get_feature(prop_market['Description - Clean'], airspace_words)

prop_market['Airspace'].value_counts()

#### Airspace Option

Words denoting options are assumed to have a common thread between referencing garages and airspace.

# Get dummy for optional airspace
airspace_opt_dummy = []
for i in prop_market['Description - Clean']:
    # Loop through features to check for 'garage' reference
    if len(set(i).intersection(airspace_words)) == 0:
        airspace_opt_dummy.append(0)
    else:
        # Check for reference of garage as option
        if len(set(i).intersection(opt_words)) == 0:
            airspace_opt_dummy.append(0)
        else:
            airspace_opt_dummy.append(1)

# Assign Dummy to Dataframe
prop_market['Optional Airspace'] = airspace_opt_dummy

prop_market['Optional Airspace'].value_counts()

# Correct overlap between dummy for airspace and optional airspace
prop_market['Airspace'] = np.where((prop_market['Optional Airspace'] == 1), 0, prop_market['Airspace'])
prop_market['Airspace'].value_counts() #confirmed, correction worked

#### Furnished

# Find words relating to furnished properties
find_words(prop_market['Description - Clean'], 'furnished', 'in')

# Get Dummy for Furnished Properties
furnished_words = ['100sqmfurnished', 'doorsunfurnishedtotal', 'furnished', 'furnished1', 'furnishedcall',
                   'furnishedcontact', 'furnishedoptional', 'furnishedref', 'furnishedsenglea', 'furnishedserved',
                   'refurnished']
prop_market['Furnished'] = get_feature(prop_market['Description - Clean'], furnished_words)

prop_market['Furnished'].value_counts()

#### Finished

# Find words relating to finished properties
find_words(prop_market['Description - Clean'], 'finished', 'in')

# Get Dummy for Finished Properties
finished_words = find_words(prop_market['Description - Clean'], 'finished', 'in')
prop_market['Finished'] = get_feature(prop_market['Description - Clean'], finished_words)

prop_market['Finished'].value_counts()

# Correct for crossovers between furnished and finished properties
prop_market['Finished'] = np.where((prop_market['Furnished'] == 1), 0, prop_market['Finished'])
prop_market['Finished'].value_counts() #confirmed, correction worked

<b>NOTE:</b> 3,340 properties are either furnished or finished, the remaining properties are assumed to be sold as shell. Dummy not created for shell properties given that 1 of the referenced features would have to be dropped anyway when modelling to avoid the dummy variable trap.

## Save Final Dataframe

# Remove Cleaned Text Column
prop_market = prop_market.drop(['Description - Clean'], axis=1)

print('Dataframe Size :', prop_market.shape)
prop_market.head()

# Save Data
prop_market.to_csv('prop_market.csv', index=False)
