==> 4 Property Market - Price Estimate.ipynb <==
# Price Estimate
## Property Market

# Adjust Notebook Display
from IPython.display import display, HTML
display(HTML("<style>.container { width:90% !important; }</style>"))

# Required Packages
import pandas as pd
import numpy as np
import statsmodels.api as sm
import math

# Load DataFrame
prop_market = pd.read_csv('prop_market.csv')
print('DataFrame Size : ', prop_market.shape)
prop_market.head()

# Double Check Duplicates
prop_market = prop_market.drop_duplicates(subset=['Links'])
print('DataFrame Size : ', prop_market.shape) # Confirmed, no duplicates

#### Roof Access Adjustment

# Display Available Property Types
prop_market['Property Type'].unique()

# OG Roof Access Variable
prop_market['Roof Access'].value_counts()

# Assign Roof Access Dummy to Properties with Automatic Roof Access
with_roof = ['Town House', 'Palazzo', 'House of Character', 'Terraced House', 'Bungalow', 'Villa', 'House', 'Semi-Detached House', 'Duplex',
             'Semi-Detached Bungalow', 'Farm House', 'Semi-Detached Villa', 'Detached Villa', 'Village House', 'Detached Bungalow', 'Guest House']
for i in with_roof:
    prop_market['Roof Access'] = np.where((prop_market['Property Type'] == i), 1, prop_market['Roof Access'])
    
prop_market['Roof Access'].value_counts()

#### Beds and Baths Adjustment to Categories

# Categories for Beds
prop_market['Beds (2)'] = np.where((prop_market['Beds'] == 2), 1, 0)
prop_market['Beds (3)'] = np.where((prop_market['Beds'] == 3), 1, 0)
prop_market['Beds (4+)'] = np.where((prop_market['Beds'] >= 4), 1, 0)

# Categories for Baths
prop_market['Baths (2)'] = np.where((prop_market['Baths'] == 2), 1, 0)
prop_market['Baths (3)'] = np.where((prop_market['Baths'] == 3), 1, 0)
prop_market['Baths (4+)'] = np.where((prop_market['Baths'] >= 4), 1, 0)

## Regression Exercise

# Get Dependent Variable - Log Price
log_price = np.log(prop_market['Price'])

# Remove Columns that are not necessary for modelling
X = prop_market.drop(['Links', 'Price', 'Description', 'Beds', 'Baths'], axis=1)

# Get Dummies for Categorical Variables
X = pd.get_dummies(data=X, columns=['Property Type', 'Locality'], drop_first=True)
print('Number of Regressors :', X.shape[1])

# Estimate Simple Linear Regression
X = sm.add_constant(X)
results = sm.OLS(log_price, X).fit()
print(results.summary())

The regression is imperfect, with several insignificant variables (determined by p-value of less than 0.05) and a large JB (with P(JB) less than 0.05, indicates non-normal residuals), but figures such as the Adj. R Squared and F-Statistic indicate that the regression is able to significantly explain most price changes (log(Price)). The regression is deemed adequate for a simple and quick property price estimate, but the model is admittedly imperfect. The regression might require heteroskedasticity robustness subject to tests for heteroskedasticity.

# Price Estimate for Property of Interest
EPrice = math.exp(results.params['const'] + results.params['Beds (3)'] + results.params['Baths (2)'] + results.params['Lift'] +
                  results.params['Pool'] + results.params['Roof Access'] + results.params['Furnished'] + results.params['Property Type_Penthouse'] + 
                  results.params['Locality_Mosta'])
print('Expected Price for Property of Interest :', round(EPrice))

# Price Estimate for Property of Interest - Excluding Pool
EPrice = math.exp(results.params['const'] + results.params['Beds (3)'] + results.params['Baths (2)'] + results.params['Lift'] + 
                  results.params['Roof Access'] + results.params['Furnished'] + results.params['Property Type_Penthouse'] + 
                  results.params['Locality_Mosta'])
print('Expected Price for Property of Interest, Excluding Pool:', round(EPrice))
